Add Layer2Rating to config
