+++ packages/config/src/layer2s/types/Layer2Rating.ts
@@ -1,0 +1,18 @@
+export type RatingCategory = 'A' | 'B' | 'C' | 'D' | '-'
+export type RatingModifier = '--' | '-' | '+' | '++'
+
+
+export interface Layer2Rating {
+  category: {
+    score: RatingCategory
+    requirements: string[]
+  }
+  modifier?: {
+    score: RatingModifier
+    items: string[]
+  }
+  thingsToImprove?: {
+    improvedScore: RatingCategory
+    requirements: string[]
+  }
+}
