+++ packages/config/src/layer2s/types/index.ts
@@ -1,5 +1,6 @@
 export * from './Layer2'
 export * from './Layer2Event'
+export * from './Layer2Rating'
 export * from './Layer2RiskView'
 export * from './Layer2Technology'
 export * from './Layer2TransactionApi'
